gitExtension: commiting changes from file explorer
Unfortunately, making changes to file structure from inside JupyterLab can be very frustrating at times. Renaming and deleting folders is not very doable because JupyterLab thinks that something is still open (probably ipynb checkpoints) that is meaningful to the user. When these changes have to be made, I use the file explorer to make them. These are those changes.