gitExtension: Added License for gitExample
The gitExample.ipynb in this folder is from Kaggle. As such, it is my understanding that it is licensed under Apache 2.0. Though it is not fully clear that this is a Derivative Work of Kaggle API, the license has been included as a precautionary measure. It is not the project license, but it is the license that most likely applies to the Kaggle notebook used.